+++ src/components/channels/Index.vue
@@ -171,6 +171,7 @@
       }
     },
     created () {
+      let activeFromLocaleStorage = LocalStorage.get.item('channels')
       this.$store.dispatch('getItems', 'channels')
         .then(() => { this.$store.dispatch('getCustomer') })
         .then(() => {
@@ -182,6 +183,9 @@
             else {
               this.active = null
             }
+          }
+          else if (activeFromLocaleStorage && this.items.filter(item => item.id === activeFromLocaleStorage).length) {
+            this.active = activeFromLocaleStorage
           }
         })
     },
@@ -222,7 +226,10 @@
         }
       }
     },
-    components: { logs, messages, QToolbar, QSelect, QInput, QIcon, QBtn, QPopover, QList, QItem, QItemMain, QItemSide, QItemTile, QTooltip }
+    components: { logs, messages, QToolbar, QSelect, QInput, QIcon, QBtn, QPopover, QList, QItem, QItemMain, QItemSide, QItemTile, QTooltip },
+    destroyed () {
+      LocalStorage.set('channels', this.active)
+    }
   }
 </script>
 <style>